requirements.txt
backcall==0.1.0
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.3.2
Flask==1.0.2
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-speech==0.36.3
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.19.0
idna==2.8
ipython==7.3.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
MarkupSafe==1.1.1
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.6.1
ptyprocess==0.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
Pygments==2.3.1
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
